improve documentation
A better recommendation for how to handle Webhook URLs.  The GitHub Webhook request has a User-Agent header set that contains GitHub-Hookshot.  If the mount configuration matches on all requests except those that match the GitHub-Hookshot User-Agent then the Reload plugin will be able to function as expected and even allow the Webhook URL to have the same Host specified as one would expect.

This is especially convenient when toadfarm is running behind nginx and you don't want to have to specify a special server name specifically for webhooking.