mojolicious v9.0 compatibility
various cleanups have been made to the routing methods provided by
Mojolicious::Routes::Route[1].

    * detour method has been removed, using ->partial(1)->to() instead
    * route method has been removed, using any instead.
    * over method has been renamed to requires.
    * via method has been renamed to methods.

the `plugins` key in a configuration file will now be used to load
deployment specific plugins. this means this top-level key is now
reserved in configuration and should not be used for other purposes.
because of this the `plugins` key in toadfarm config has been renamed to
`tf_plugins`

the minimum required version of Mojolicious has been bumped to 8.73,
which is the version before 9.00 - this allows users to upgrade this
distribution without having to go fully v9 of Mojolicious

[1] https://github.com/mojolicious/mojo/wiki/Upgrading